% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_plot.R
\name{ciplotter}
\alias{ciplotter}
\title{Plot confidence intervals}
\usage{
ciplotter(cix, ciy.lo, ciy.hi, eps, ...)
}
\arguments{
\item{cix}{values of the x axis}

\item{ciy.lo}{spread of confidence intervals}

\item{ciy.hi}{spread of confidence intervals}

\item{eps}{width of the whiskars}

\item{...}{additional base plot arguments}
}
\description{
Used by \code{\link{scenario_plot}}.
}

