% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_plot.R
\encoding{utf-8}
\name{scenario_plot}
\alias{scenario_plot}
\alias{scenario_plot80}
\title{Plot scenario simulation results}
\usage{
scenario_plot(dat, sizes, n_sim, to_n, tn_n = 8 - to_n, multi = 1,
  jitter = 0.05, letter = "", ...)

scenario_plot80(dat, sizes, n_sim, multi = 1, letter = "", ...)
}
\arguments{
\item{dat}{simulated data}

\item{sizes}{An array of sample sizes to be simulated. Can be single value.}

\item{n_sim}{number of simulations. 1000 is a start, 10000 was used in paper, but takes a long time}

\item{to_n}{Number of indicators in a Trait relating to Outcome}

\item{tn_n}{Number of indicators in a Trait Not relating to outcome.}

\item{multi}{influences cex of certain plot variables.}

\item{jitter}{Avoid overalp between lines}

\item{letter}{assigns plot a letter, useful for combining multiple plots}

\item{...}{additional options for axis}
}
\description{
Plots the results like in Study 1 of the Vainik et al. paper an overview of the indicator exclusion results. Starred indicators are excluded in the indicator exclusion procedure. See \code{\link{scenario_sim}} for details. NB! Scenario with no ION violations needs \emph{scenario_plot80()}
See  \code{\link{scenario_sim}} for examples
}
\section{Functions}{
\itemize{
\item \code{scenario_plot80}: For plotting scenarios where ION is not violated
}}

